Move printing utilities into GitRepo
