+++ clowder/clowder_repo.py
@@ -10,7 +10,6 @@ import sys
 from git import GitError
 from termcolor import colored, cprint
 
-import clowder.git.printing as git_print
 from clowder.git.repo import GitRepo
 from clowder.utility.connectivity import is_offline
 from clowder.utility.execute import execute_command
@@ -37,7 +36,8 @@ class ClowderRepo(object):
             sys.exit(1)
         except (KeyboardInterrupt, SystemExit):
             sys.exit(1)
-        git_print.status(self.clowder_path)
+        else:
+            GitRepo.status(self.clowder_path)
 
     def branches(self):
         """Return current local branches"""
@@ -49,7 +49,7 @@ class ClowderRepo(object):
         clowder = GitRepo(self.clowder_path)
         if self.is_dirty():
             print(' - Dirty repo. Please stash, commit, or discard your changes')
-            git_print.status(self.clowder_path)
+            GitRepo.status(self.clowder_path)
         else:
             clowder.checkout(ref)
 
@@ -124,8 +124,8 @@ class ClowderRepo(object):
             print(' - Fetch upstream changes for clowder repo')
             clowder = GitRepo(self.clowder_path)
             clowder.fetch('origin')
-        project_output = git_print.format_project_string(repo_path, '.clowder')
-        current_ref_output = git_print.format_project_ref_string(repo_path)
+        project_output = GitRepo.format_project_string(repo_path, '.clowder')
+        current_ref_output = GitRepo.format_project_ref_string(repo_path)
 
         clowder_symlink = os.path.join(self.root_directory, 'clowder.yaml')
         if not os.path.islink(clowder_symlink):
@@ -181,13 +181,13 @@ class ClowderRepo(object):
 
     def git_status(self):
         """Print clowder repo git status"""
-        git_print.status(self.clowder_path)
+        GitRepo.status(self.clowder_path)
 
     def _validate_groups(self):
         """Validate status of clowder repo"""
         clowder = GitRepo(self.clowder_path)
         if not clowder.validate_repo():
-            git_print.validation(self.clowder_path)
+            GitRepo.validation(self.clowder_path)
             print()
             sys.exit(1)
