+++ clowder/model/fork.py
@@ -6,7 +6,6 @@ import os
 
 from termcolor import cprint
 
-import clowder.git.printing as git_print
 from clowder.git.repo import GitRepo
 
 
@@ -33,6 +32,6 @@ class Fork(object):
         if not GitRepo.existing_git_repository(self.path):
             cprint(self.path, 'green')
             return
-        project_output = git_print.format_project_string(self.path, self.path)
-        current_ref_output = git_print.format_project_ref_string(self.full_path())
+        project_output = GitRepo.format_project_string(self.path, self.path)
+        current_ref_output = GitRepo.format_project_ref_string(self.full_path())
         print(project_output + ' ' + current_ref_output)
